--- RadixWallet/Features/DApps/DAppsDirectory+AllDapps.swift
@@ -1,27 +1,0 @@
-// MARK: - AllDapps
-extension DAppsDirectory {
-	@Reducer
-	struct AllDapps: Sendable, FeatureReducer {
-		@ObservableState
-		struct State: Sendable, Hashable {
-			init() {}
-		}
-
-		typealias Action = FeatureAction<Self>
-
-		enum ViewAction: Sendable, Equatable {
-			case appeared
-		}
-
-		var body: some ReducerOf<Self> {
-			Reduce(core)
-		}
-
-		func reduce(into state: inout State, viewAction: ViewAction) -> Effect<Action> {
-			switch viewAction {
-			case .appeared:
-				.none
-			}
-		}
-	}
-}
